ref: move all task and project dto to related repository
