+++ src/main/java/com/lab/project_tracker/dto/project/ProjectDto.java
@@ -1,6 +1,9 @@
-package com.lab.project_tracker.dto;
+package com.lab.project_tracker.dto.project;
 
 import com.lab.project_tracker.util.ProjectStatus;
+import jakarta.persistence.Column;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,13 +12,13 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-/* this is dto is used project is nested */
+/* this is dto is used for inserting or updating */
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ProjectDtoSummary {
-    private Long id;
+public class ProjectDto {
     private String name;
     private String description;
     private LocalDate deadline;
