+++ src/main/java/com/lab/project_tracker/dto/task/TaskDto.java
@@ -1,4 +1,4 @@
-package com.lab.project_tracker.dto;
+package com.lab.project_tracker.dto.task;
 
 import com.lab.project_tracker.util.TaskStatus;
 import lombok.AllArgsConstructor;
@@ -8,15 +8,15 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 
-/* this is dto is used task is nested */
+/* this is dto is used for inserting or updating */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class TaskDtoSummary {
-    private Long id;
+public class TaskDto {
     private String title;
     private String description;
     private LocalDate dueDate;
     private TaskStatus status;
+    private Long projectId;
 }
