+++ src/main/java/com/lab/project_tracker/service/TaskService.java
@@ -1,10 +1,8 @@
 package com.lab.project_tracker.service;
 
 
-import com.lab.project_tracker.dto.ProjectDto;
-import com.lab.project_tracker.dto.TaskDto;
-import com.lab.project_tracker.dto.TaskResponseDto;
-import com.lab.project_tracker.model.Project;
+import com.lab.project_tracker.dto.task.TaskDto;
+import com.lab.project_tracker.dto.task.TaskResponseDto;
 import com.lab.project_tracker.model.TaskEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
