+++ src/main/java/com/lab/project_tracker/service/impl/TaskServiceImpl.java
@@ -1,7 +1,7 @@
 package com.lab.project_tracker.service.impl;
 
-import com.lab.project_tracker.dto.TaskDto;
-import com.lab.project_tracker.dto.TaskResponseDto;
+import com.lab.project_tracker.dto.task.TaskDto;
+import com.lab.project_tracker.dto.task.TaskResponseDto;
 import com.lab.project_tracker.exception.ProjectNotFoundException;
 import com.lab.project_tracker.exception.TaskExistsException;
 import com.lab.project_tracker.exception.TaskNotFoundException;
